test: increase timeout and poll interval for slow test
In `VariableSearchTermsQueryChunkIT.shouldReturnAllMatchedUserTasks`, the Awaitility check
was failing after 1 minute, even though the app logs showed the count of returned records was
consistently increasing. The last observed response before failure contained 48 items.

This suggests that the data was still being indexed, but the test didn't allow enough time.

To prevent false negatives and flakiness, the timeout was increased to 2 minutes and
the poll interval to 5 seconds.