fix: partition bootstrapping unable to complete when snapshot exists
When a snapshot is present in the store, check if it's a snapshot for
bootstrap and delete it if it is.
Then fetch again a snapshot, in case the PartitionBootstrapOperation was
canceled and recreated.
If the id is not from a snapshot for bootstrap, then fail the start step
as manual intervention is required