feat: use service exception in REST layer
Replaces the in-depth error mapping in the REST layer with a lightweight mapping to REST DTOs.
The REST layer maps the status of the service exception to an HTTP status and reuses the exception message as detail.
No further logs are written in the REST layer for errors, except when anything else than a service exception occurs.

The document controller removes any special mapping of document errors. This happens in the service layer already.