fix: apply checkpoint confirmation state change
Unlike engine processors, we don't have a result builder method to
append the event and apply it at the same time. By only appending the
event, we updated the state on replaying followers, but not on the
leader. We fix this by copying the approach from
`CheckPointCreateProcessor`: just applying the state change inline.
To avoid future bugs, we reuse the whole event applier logic.