feat: support ad-hoc subprocess job result
This adds 2 properties to the JobResult. One is for setting a type. Depending on the type of the JobResult a different
set of properties will be set on the record. The supported types at this time are the user task, and the ad-hoc
subprocess.

The user task properties all exists already. For the ad-hoc subprocess we need to introduce a new one. This is the
activateElements property. This property will be used to determine which elements of an ad-hoc subprocess need to be
activated, and what variables should be created.