feat: [BREAKING CHANGE] add new exporter state in cluster configuration
The state "CONFIG_NOT_FOUND" indicates the exporter is still active, but
since the configuration is not found it might not be possible to export to
them anymore.

Breaking Change: Renamed the enum in ExportState in `topology.proto`. This is
not a breaking change in protobuf as the binary representation is not changed.
Hence this is not technically a breaking change.