feat: [BREAKING CHANGE] add new exporter state in cluster configuration
The state "CONFIG_NOT_FOUND" indicates the exporter is still active, but
since the configuration is not found it might not be possible to export to
them anymore.

Breaking Change: Renamed the enum in ExportState in `topology.proto`. This is
not a breaking change in protobuf as the binary representation is not changed.
Hence this is not technically a breaking change.

(cherry picked from commit 35af30c96712532e75e4a4721a2ee7e0e30e367f)
(cherry picked from commit 73ed0438acd3e86dc02d253d6dea5cfa9b195b7a)