feat: summarize compensation subscription
This adds a summary for compensation subscriptions.

The summary is different for compensation subscriptions that have not
yet been triggered, and those that have.

In case the compensation subscription has not yet been triggered the
summary describes which compensation handler is registered for the
compensable activity. This makes it clear what would happen when the
compensation is triggered, i.e. the compensation handler is triggered to
compensate the compensable activity.

In case the compensation subscription has been triggered, the summary
describes which event triggered it, and what effect it had. i.e.
executing the compensation handler. Only in this case can there be
variables present.

It's not useful to present the process information again on the events
where the subscription has already been triggered. It's easy to infer
from the other keys in use.