refactor: extract task headers builder interface
The following elements (currently) can have task headers specified
according to the bpmn-model API:
- job worker tasks (like service task, script task, send task, etc)
- user task (for job worker implementation)
- throw event (for job worker implementation of message throw events)

The method schema (name and parameters) is the same for all of these, so
a new interface is easily extracted. This is especially useful for
process generation using the model api, e.g. as needed in parameterized
tests.

This is not a breaking change.