test: ensure task headers access in exec listeners
Adds a parameterized test case that ensures we add task headers to the
START and END execution listeners. Specifically as custom headers on the
created job. We also verify that the headers remain on the COMPLETED
event of the job, although that is less relevant.

The test assumes that the specific scenario actually supports setting
task headers by pre-applying the task configurer. The result should be
an AbstractTaskBuilder that implements BuilderWithTaskHeaders. If that's
not the case, there's no reason to run this test because we can't set
task headers on this task. Perhaps in the future, we'll allow specifying
task headers for all tasks to accommodate this.

The test itself is straightforward:
- run the process instance and complete the start execution listener
- complete the task and complete the end execution listener
- now verify the created and completed job records for custom headers