refactor: check if index exists before migration
If the index does not exist, the application exists with an error. It
might not happen in production because the indices are always created
by Operate and Tasklist. However, it is a good practice to check it
anyways.

This check is also required by a test that verifies if the docker container
starts. In that test we do not have to fill up ES/OS with data.