feat: add new StackdriverLayout JSON schema
Adds a new StackdriverLayout.json template which can be used by Log4j2's
built in JsonTemplateLayout. This will allow us to completely remove our
old Stackdriver layout, and make better use of garbage-free logging.

The commit also adds including the Spring application version via
resource filtering, since this can then be used to have as the default
service version.

This should be split into two commits, but as it's quite small, the
overhead of doing so is likely more than it costs the reviewer.