ci: remove has-no-project condition from add-to-projects
With the new issue template, labels are set with a dedicated workflow `labeler.yml` after creation,
additionally we have the `add-to-quality-board` step that will, on initial creation, add a bug to the Quality board.

This broke adding issues to project boards like CamundaEx like this one https://github.com/camunda/camunda/issues/35630
which I had to add manually.

I don't see a need for the condition to be used still.
It was originally added to avoid that issues already added to the ZPA project but labeled with component/zeebe
to get added to the ZDP board. Now with the mono-repo we have rather distinct component labels and if something is
labeled with a component or even multiple, I would argue it should always end up on respective projects.