test: align remaining 8.6 appliers
We discussed how to deal with the remaining changes. Even if we cannot
host them, they are already available in released versions. In an effort
to take profit from these golden files, we can simply accept them for
now and merge the changes. Afterward the test will help defend against
making further changes to existing appliers.