feat: checksum change is enough to take a new snapshot
This is the core change to allow taking new snapshots even if none of
the positions changed. We now include the combined checksum in snapshot
id and directory.
Because the checksum is only known after taking the snapshot, we can
only move the snapshot to the correct directory at the end.