test: add element instance filter DI QA tests
Adds test cases for advanced search filtering on element instance key in decision instance acceptance tests.
Adds test cases for process instance key filtering as well since this was previously not covered.
The test class deploys a process with a business rule task for this to have a process instance and element instance key.