fix: use 204 no-content responses instead of 202
202 Accepted is meant to be used for asynchronous processing, where the
server has accepted the request and will process it.

204 No Content is meant to be used for synchronous processing, where the
server has already processed the request and there is simply no content
to respond.

In our case, the responses are always sent after the action has already
taken place. For example, when assigning a role to a user the response
is sent after the role was assigned to the user successfully. In other
words, we won't continue processing the request because it was already
processed. Using 204 No Content is thus a better fit for these
endpoints.