feat: define input collection behavior
This behavior will be responsible for managing the input collection. It'll evaluate the expression and ensure it gets
stored in state. It's also responsible for retrieving the collection from the state when necessary.

A future expansion will be removing the input collection from state when the MI body completes.

(cherry picked from commit b4eb19aea1111ceb990507a15886eda51e65987e)