fix: activate child using the input collection from the flowscope
This introduces a slight change in behavior. Before we would evaluate the expression using the child context. This
changes it so that it might be evaluated (for running instances) using the flow scope context. As the input is first
evaluated using the MI body (flow scope here) it makes sense to keep doing that for the children as well. We don't
support changing the input collection. If anything using the child context here creates a bigger risk of the input
collection changing.

Famous last words: This should not introduce any bugs.