fix: schedule jobs stream recreation
Previously the streamingTimeout was used as request timeout on the job stream control,
this lead to unnecessary warning logs when the timeout hit and the stream got closed to be recreated.

With this, the stream control is scheduled to get cancelled and immediatly recreated with only debug log statements.

(cherry picked from commit 0f5b4533517c55555d5b53bf3fdc89f78b780bdb)