feat: add authorization scope wrapper for resource ids
Add an AuthorizationScope wrapper class for resource ids.
The wrapper class is contained in the security-core module
and can be re-used by the Search layer and the Workflow
Engine.

The goal of the wrapper class is to provide contextual
information on the resource IDs, i.e. if an ID is a
WILDCARD character allowing a certain permission on all
resources, or if it is a specific String-based identifier.