refactor: adjust the service layer authorization checks
Use the new AuthorizationScope class in the service layer AuthorizationChecker methods.

When resource IDs are retrieved from secondary storage, they are wrapped extracted from the Authorization Entity and wrapped in an AuthorizationScope
instance.

The wrapper is only mandatory when resource IDs enter
the authorization check code, and are normalized when
they exit it.