test: Stabilize flaky tests
Some CPT tests fail frequently in the CI builds because the assertion timeout is too low.

Harmonize the assertion timeout for all CPT tests by introducing a new JUnit extension that sets the timeout instead of
the individual tests.

For local development, use a timeout of 1 second. For CI, set the timeout to 10 seconds via Maven plugin configuration.