test: ensure delete user cannot be found by key
A user can be found by username and by key. When deleting a user we
should ensure that both are verified.

Note that this test passes even though the userKeyByUsernameColumnFamily
is not cleaned up. This is because that column family only stores a
foreign key, not the entire object so the result is still empty.