test: clear user from all column families
This test goes deeper than the existing
shouldNotFindUserByUsernameAfterDelete and
shouldNotFindUserByKeyAfterDelete tests by verifying the state rather
than what the state's API offers.

This was necessary as we forgot to remove the user from all column
families. I tried to think about ways to test this for future column
family additions, but that's harder to solve. Perhaps we could do
something with the foreign key concept or so, but I don't see it now. So
I think this is the best we can do for now.