fix: on user deleted also delete user by key
We should delete a user from both the userByUsername and the
userKeyByUsername column families. The latter has a foreign key to the
user, so we should delete it before deleting the user by username.

Taking this approach rather than deleting it from both column families
inside the state class ensures that it's clear what the event applier
does.

Additionally, neither method has access to both the username and the
user key, so implementing it as delete(username) would require a lookup,
that is now unnecessary.

See: https://github.com/camunda/camunda/issues/13813