feat: add processor for AHSP inner instance
The AHSP inner instance is a container element which wraps around the element inside the AHSP. This inner instance has
it's own element type and requires it's own processor. As of now this processor is quite simple. It has to do nothing
except ensure that element completions/terminations are passed along to the AHSP itself. Direct activation is not
supported at all.