feat: wrap AHSP element in inner instance
I've added a new behavior class. This class will take care of activating the element of an AHSP. Since we want to wrap
the elements inside of an inner instance this new behavior will first create this inner instance. It does so by writing
the activating and activated events. By writing these events we create the inner instance element instance. Using the
key of this inner instance we activate the elements itself. We must ensure the flowscope key becomes the key of the
inner instance. In turn the inner instance has the key of the AHSP as the flowscope key.