test: update golden files
The MultiInstance INPUT_COLLECTION_EVALUATED is a new event applier. We need to add the golden file.

The ELEMENT_COMPLETED and ELEMENT_TERMINATED event appliers have indeed changed.
They will now delete an input collection from sate, if it exists!
Because we use deleteIfExists() this is fully backwards compatible we can safely
add it without requiring a new version of the event appliers.