fix: fail if ownerTypesToOwnerIds is not valid
The semantic of AuthorizationFilter#ownerTypeToOwnerIds is not clear, meaning, it can happen
that a owner type is provided but its value is null or empty. In that case, the query would
randomly return an authorization matching the owner type. But it would not consider the owner ids
in that case.

To avoid any unexpected behavior, for example granting users randomly with access to resource,
throw an exception if the ownerTypeToOwnerIds map contains type without any owner ids.