feat: restore from multiple contiguous backups
This allows users to specify a list of backup ids. When these are
contiguous, i.e. when the user has enabled this by setting
`zeebe.broker.experimental.continuousBackups` to true, restore will:
1. Pick the snapshot from the first backup
2. Rewrite the journal

We rewrite the journal by reading from each backup journal and appending
to the target journal. The index of the target journal is reset to match
the first index of the first backup. Only records in the range of
`(previousBackup.checkpointPosition, currentBackup.checkpointPosition]`
are read. The result is a contiguous journal without any overlaps.

We still support restore from a single backup because the behavior is
the same: we still pick the snapshot and effectively truncate the log
to the checkpoint position.