feat: toggle recording exporter reset for identity
When using EngineRule.withIdentitySetup(), the IdentitySetup Initialize
command is executed which results in many records on the log stream. In
some cases, it may be useful to access these records from the recording
exporter.

These changes add a toggle that enables skipping the reset of records
after the identity setup has been initialized.