test: ensure default admin role is created
This tests the properties of the default admin role's authorizations.

The test is setup using some uncommon usages related to the
RecordingExporter that help in preserving the recorded records for
individual test cases while only starting the engine once. This helps
speed up the testcases. These are:
- ENGINE resets the recording exporter ONLY_BEFORE_AND_AFTER_ALL_TESTS
  to avoid side effects to other tests
- ENGINE additionally uses NO_RESET_AFTER_IDENTITY_SETUP

The assertions themselves are structured using
satisfiesExactlyInAnyOrder to ensure that adding a new authorization to
the default role will fail the test, without the need for an explicit
size assertion.

The common properties between all authorizations can be tested in one go
using allSatisfy.