feat: remove old sequence flow migration applier
We've replaced the previous sequence flow migration with the new
deletion and re-creation logic. We no longer need the existing applier.
Normally, we should not remove any appliers to provide backward
compatibility and prevent versioning errors during cluster upgrade.
However, this applier was only available in alpha versions and not
released at all, so it is safe to remove.

Tests were already updated to conform with deletion and creation logic.
Therefore, no test changes are needed.