fix: only update exported position after flush
When skipping a record, we must not yet update the exported position
because we might still have unflushed records that we can't acknowledge
yet.

Instead, we wait on the next flush: either a scheduled one or when
exporting the next record.

(cherry picked from commit 0a33298b6950a1a35acdd43fe26ab89115ee4809)