style: return switch expression result
Checkstyle failed with the message that the switch was missing a default
branch. But there really is no need for a default branch here. Instead,
we can yield 'this' in each case to allow returning the result of the
switch expression. With that the java compiler is able to detect missing
cases resulting in a compile error.

It's a few more lines of code, but the result is easier to maintain.