feat: add id regex config property
Allow to configure the identifier regex as part of
the InitializationConfiguration properties. The
property is placed in the initialization config
as the ID regex should be set when the system is
set up. Otherwise, invalid IDs may be entered in
the system.

A guard validation to prevent the allowance of
the asterisk character is added as well. The
asterisk is a reserved character, used for
all-inclusive access to resources.