fix: allow update from leaders with lower term after partition bootstrap
Previously, it was not possible to overwrite a partition leader with lower
term. This is required to handle out-of-order membership events. With the purge
operation, a partition is re-bootstrapped with a clean state. As a result the new
leader will start with term 1.

To handle out-of-order events properly, instead of updating the existing in-memory
cluster state, we always built it from the set of received member properties. This
way we generate a new topology from the latest member properties of all brokers that
are known to this client/gateway. We still use the term for conflict resolution to determine who is the latest leader. But since the old leader do not exist anymore,
the newly generate topology will have the correct leader after the partition
re-bootstrap.

This approach consumes extra memory to keep track of all memberProperties in
BrokerTopologyManager. In addition, it has to iterate over all properties to generate
the new topology. However, membership updates are not very often to optimize this path.
So this is an acceptable overhead.