feat: introduce new process instance permissions
Currently cancelling and modifying a process instance use the
UPDATE_PROCESS_INSTANCE permission on the PROCESS DEFINITION resource.

This is problematic because:

- It's a common use case that operators should be able to adjust a
  process instance, but not quit it
- It's a common use case that operators should be able to cancel a
  process instance (either through cancel API or modification), but not
  delete it entirely from the secondary storage (unlike
  DELETE_PROCESS_INSTANCE permission) Please enter the commit message
  for your changes. Lines starting

MODIFY_PROCESS_INSTANCE is a dedicated permission because it's more than
either CANCEL_PROCESS_INSTANCE or UPDATE_PROCESS_INSTANCE. An
alternative would be to check both of these permissions, but we prefer
the dedicated MODIFY_PROCESS_INSTANCE as that mirrors the batch
permissions better and may be easier to understand for users.

So, now we can introduce CANCEL_PROCESS_INSTANCE and
MODIFY_PROCESS_INSTANCE as a new permissions.