feat: support reset before all and after each test
This option was still missing but is actually really handy when the
EngineRule is used as a @Rule (per test) rather than as a @ClassRule
(once for all tests).

In the case that the engine rule is set up for each test, then we
actually may not want to reset before each test because we may want to
include the identity setup records that are added before the test run.

Additionally, in such a mode we do still want to ensure that the tests
are isolated. We can achieve that by resetting after each test. To
additionally, isolate the entire test class from other test classes, we
can reset before all tests as well.