feat: require mode when no reset after identity
When the NO_RESET_AFTER_IDENTITY_SETUP mode is used, it does not make
sense to use the BEFORE_EACH_TEST mode for the EngineRule's
recordingExporterTestWatcher because the recording exporter would still
be reset before the test case (after the identity setup).

To avoid confusion, we can simply require that the
ResetRecordingExporterTestWatcherMode is set on the EngineRule to
something that is not BEFORE_EACH_TEST when
NO_RESET_AFTER_IDENTITY_SETUP is requested.

The instructions clarify what options are possible.