feat: complete the ad-hoc sub-process if the completion condition is fulfilled
When a Job is completed and the completion condition is fulfilled we directly try to complete the ad-hoc sub-process.
However, it could be that there is still running children. If there is we cannot complete the ad-hoc sub-process just
yet. We must wait until these inner instance are either completed or terminated. We can catch these cases in the
afterExecutionPathCompleted method. This is called when a child is completed or terminated. Since we have stored the
completionConditionFulfilled flag on the element instance we can access this and try to complete the ad-hoc sub-process
accordingly.