fix: print compact log on failure
The RecordingExporterTestWatcher inside the EngineRule can print the
compact record log ONLY IF the EngineRule is set up as `@Rule`.

If the EngineRule is set up as `@ClassRule`, the internal
RecordingExporterTestWatcher's failure method is not called at all.

To overcome this, tests that specify EngineRule as `@ClassRule` need to
additionally add the `@Rule` RecordingExporterTestWatcher themselves.

To avoid this additional RecordingExporterTestWatcher resetting the
logstream when starting the tests, we need to set its reset mode to
NEVER.

The outcome is that:
- EngineRule's RecordingExporterTestWatcher resets the RecordingExporter
  before all tests and after all tests
- EngineRule does not reset the RecordingExporter after the identity
  setup
- The test class's RecordingExporterTestWatcher does not reset the
  RecordingExporter at all, but can print the compact log

Note that JUnit 5 would offer better ways to compose these concepts. The
above is simply a JUnit 4 limitation.