feat: merge user RocksDB config with our defaults
This changes behavior in how we deal with user-provided RocksDB config.
Instead of only using user-provided config if any is set, we now merge
it on top of our own default config.
This gives users much more flexibility to configure RocksDB and benefit
from our default configuration at the same time.

Because RocksDB does not expose a method to merge database or column
family options directly, we now define our default config as properties
and merge them with user provided properties before letting RocksDB
translate them to their own format.
Unfortunately, that means we have to deal with https://github.com/facebook/rocksdb/issues/13841.
We do this by calling out to libc `sprintf` to format any `double`s
exactly how RocksDB would do it.
Because the configuration via properties is not as safe as the
programmatic approach used before, we add a golden file test that
verifies that the actual RocksDB options are exactly what we expect.