fix: filter out NONE_INSTANCE contexts and fix NPE
Before this fix, if a regular variable e.g. `my_var` was used in a FEEL expression,
the `ScopedEvaluationContext.none()` instance was returned by `CombinedEvaluationContext#getVariable`
that led to the inability to retrieve the `my_var` variable from storage.
By filtering it out, we can proceed to `VariableStateEvaluationContext` which can retrieve the variable from storage.

Also this commit resolves NPE exceptions in `GlobalVariableEvaluationContext` and
`VariableStateEvaluationContext` as `variableState.getVariable(...)` returns `null` when variable not found.