refactor: simplify fromAi() function extraction by allowing only constant values
Previously, it was possible to specify something like to dynamically
build the schema or the description.

```
=fromAi(
  toolCall.myValue,
  string join(["A", "simple", "value"], " "),
  "str" + "ing",
  context put({}, "enum", ["A", "B", "C"]),
  { optional: not(false) }
)
```

As this might be misleading due to the different evaluation context and
it makes the implementation more complex, this now changes the
evaluation to only support constant values (e.g. ConstContext) as
function invocation params.