test: keep identity setup under 32 KB
This adds a test case to ensure that the identity setup commands are
kept within a limit of 32 KB to ensure it can be used in combination
with low max message size configurations.

Currently, the test fails because the identity setup INITIALIZE command
is processed using command batching. The follow-up commands and events
together exceed 32 KB of data. This results in a warning log and a
rollback of the processing after which processing is attempted again
without command batching.

As this is slows down startup, we want to avoid this.

The test works by waiting for the initialization to complete and then
verifying that the specific warning was not logged. If the warning log
is ever changed, this test becomes useless so we add a comment to the
specific log message to update this test case.