fix: add a wrapper for `JwtDecoder`
Same as for the previous wrapper,the main purpose of this is to ensure `JwtDecoder`
is created rather on a startup or later during retries:

- on a startup it will attempt to create a decoder
- if this fails, it'll create `SafeInitProxy` instance (see previous commits) and pass 2 arguments:
 a supplier to create a `JwtDecoder` and en error listener (log entry will be present)
- this will schedule async retries of creation
- startup will succeed
- created proxy will keep attempting to create the decoder in the background
- at the same time, if client makes a call (any), `decode` will attempt to get a value
from the proxy (if successful that means the IdP is available). If the value is not present
(meaning IdP is still not available) it'll throw an exception that will be returned to the client.

Open question: what the error message should be here?
JwtDecoder relies on ClientRegistrationRepository so I guess the error here should
probably be the same?