test: add the automatically created AHSP elements variable to the assertion
In https://github.com/camunda/camunda/pull/36435 we've added a variable created by the engine that provides some
metadata of the ad-hoc sub-process. This test now encounters this variable.

I chose to add it here to the assertion. Alternative would be excluding it, but I figured adding it is more explicit.