test: configure tenant access case insensitive
These tests verify that the resource access controller has tenant access
or not through the multi tenancy checks enabled property. As this
property can be configured both using camelCase and kebab-case, we can
verify the 4 combinations.

Note that enabling it using kebab-case currently fails, we'll fix this
in a following commit.