fix: support multi-tenancy.checks-enabled config
The property was only checked using multiTenancy specifically, which
meant that the kebab-case style (multi-tenancy.checks-enabled) was
ignored completely.

We can easily resolve this by dynamically providing the bean using the
SecurityConfiguration which correctly determines the config option
regardless of camelCase or kebab-case because it's based on spring's own
property loading.