ci: sign cherry-picked commits by backport-action
This adds a GPG key for backport-action with which it can sign the
cherry-picked commits that it creates.

This is achieved using crazy-max/ghaction-import-gpg and two secrets.
The secrets specify the private key and corresponding passphrase. These
are imported and the git config is adjusted to allow signing of commits
using this key. In addition, backport-action is instructed to use the
corresponding user id and email when creating git commits - which it
does only through cherry-picking.

crazy-max/ghaction-import-gpg seems to be a respectable action to import
a gpg key for usage in signing git commits. I've tested it in
korthout/backport-action-test successfully using the exact same changes
as made here. The repo seems to receive semi-regular updates, although
it's behind on its opengpg version currently. Given the context of this
workflow (used only to set up a gpg key for git commit signing by the
backport-action), this is an acceptable security risk. The exposed gpg
key (set up in a secret) is linked to the github.com/backport-action bot
account only.

See: https://github.com/crazy-max/ghaction-import-gpg
See: https://github.com/korthout/backport-action-test/pull/386