feat: prevent backups during scaling
A checkpoint is created and a failed backup is taken during scaling. A
rejection response is sent to the user.

A checkpoint must be created for consistent checkpoints across the cluster.
If not, we may end up in a state where one partition has a checkpoint before
scaling and the other one has a checkpoint after scaling resulting in a
"complete" backup which is not consistent with each other.