feat: add archunit-tests module
Adds a new module that can be used to specify ArchUnit tests for the
orchestration cluster.

These tests have access to the entire production code through a
dependency on dist (io.camunda:camunda-zeebe).

The tests need to be run by surefire rather than failsafe to ensure
correct execution and access to loaded classes. With failsafe, tests are
executed but failures are not detected somehow. I did not understand why
this is. Note that the parent module (io.camunda:camunda-qa) attempts to
run all tests with failsafe rather than surefire so the archunit-tests's
pom explicitly re-configures this.

The choice to create a new module was made as it:
- allows to easily run these tests together
- depending on dist ensures it runs on all production code, including
  modules added later without needing to change dependencies
- helps others to find a fitting place for their ArchUnit tests
- groups together ArchUnit tests for easily gaining an overview

The downsides are:
- long feedback loop - I value this less than the long term ease of
  maintenance.
- doesn't run on non-prod code - No need to solve this now.

See: https://camunda.slack.com/archives/C08F5RD5X8B/p1755527617038109