test: ensure no camel case in property conditional
The ConditionalOnProperty annotation requires kebab-case in its
attributes in order to correctly match them using relaxed binding to
spring loaded properties. In other words, to match to both my-example
and myExample, the @ConditionalOnProperty must refer to my-example. If
it refers to myExample, then it only works explicitly on myExample and
relaxed binding to my-example doesn't work.

To avoid this, we add an ArchUnit test that verifies that none of the
attributes of ConditionalOnProperty use camelCase. It does this simply
by checking for capitalized letters. If it fails, the error message
includes all the needed details, like the location of the forbidden
annotation, the specific attribute, and the failure.

To overcome the problem, simply switch to using kebab-case in the
attributes of ConditionalOnProperty.

(cherry picked from commit d67a9f84a19353e7d3c5410cb6b309a8c1d446c5)