fix: use kebab-case for operate.webapp-enabled
ConditionalOnProperty requires specifying names of properties using
kebab-case.

This ensures that camunda.operate.webapp-enabled can be used rather
than requiring camunda.operate.webappEnabled explicitly.

Note that the docs document it as camunda.operate.webappEnabled, but
this change is backward compatible, as camunda.operate.webappEnabled
can still be matched due to relaxed binding. We decided to use
kebab-case for configs in the docs, but they have not been adjusted yet.