fix: use kebab-case for tasklist.importer-enabled
ConditionalOnProperty requires specifying names of properties using
kebab-case.

This ensures that camunda.tasklist.importer-enabled can be used rather
than requiring camunda.tasklist.importerEnabled explicitly.

Note that the docs document it as camunda.tasklist.importerEnabled, but
this change is backward compatible, as camunda.tasklist.importerEnabled
can still be matched due to relaxed binding. We decided to use
kebab-case for configs in the docs, but they have not been adjusted yet.