fix: no longer persist task claim directly
This call to persistTaskClaim can conflict with the CamundaExporter
updating the same index on the `ASSIGNED` event. To avoid that race
condition, we have decided to rely exclusively on the CamundaExporter.
It is thus no longer necessary to persist the task claim directly in the
Tasklist V1 API.

This has a downside. The user must rely on eventual consistency.
Refreshing the page can temporarily show the previous value again.