refactor: reflect codebase structure in ArchTests
This change moves the RequireKebabCaseInConditionalOnPropertyArchTest to
the io.camunda package as it covers all those classes.

This makes it easier for other ArchTests to also align to the structure
of the codebase that they work with. The end goal is a module that
organizes its test in way where it's easy to see at a glance what code
is covered by these tests.