test: dont mock data objects
Mocking data objects can lead to surprising test failures. Like a
NullPointerException.

The problem here was that two commits back we changed what object was
used in the test through a simple refactoring. Rather than relying on
the returned object of persistTaskClaim, we simply called a setter on
TaskEntity that returns itself. While that method may be implemented on
TaskEntity, it is not implemented when using mock(TaskEntity.class).

In general, we should attempt to avoid mocking when possible as it
quickly leads to hard to maintain test cases. This was a great example
to showcase that.

The tests are fixed by simply constructing the data object with the
reasonable data: id and assignee.