test: check isAuthorized edge cases
This adds a unit test for AuthorizationChecker that verifies the correct
functioning of the isAuthorized method.

It uses parameterized tests for scenarios with anonymous users and
authenticated users, clients and mapping rules, where it considers
whether they can access a resource. Cases cover all reasonable
combinations of authentication, authorization, and resource to access.
An expected enum is used to easily clarify the expected outcome: Access
allowed, Access denied. Some scenarios overlap, but that's fine as this
test is cheap to maintain.

The AuthorizationChecker depends AuthorizationReader for which a fake
implementation is provided. This is done rather than using Mockito to
mock it, as using a fake leads to tests that are less dependent on the
implementation of AuthorizationChecker and how it interacts with
AuthorizationReader, but rather depends on a basic implementation of the
reader. These tests thus run fast and are cheap to maintain (they don't
depend on the implementation but on the interface).

The fake reader is implemented as an in-memory list of authorizations
that can be created so they can be accessed using getByKey and search.
The search is implemented as simple as possible, only considering filter
options that are not null. The fake is closed after all tests (clearing
the stored authorizations), which is not strictly necessary. However, I
felt this highlighted how this data store is available throughout all
tests.

Note that an alternative way to test the AuthorizationChecker would have
been to set up Elasticsearch using Testcontainers. However, I found this
to be slow and brittle. The current tests using the fake reader run in a
fraction of a second instead.