test: cover collectPermissionTypes method
Adds another ParameterizedTest that covers the collectPermissionTypes
method of AuthorizationChecker. Again we use FakeAuthorizationReader to
easily set up the available authorization entities. But different from
the IsAuthorizedTests, here we set up a new FakeAuthorizationReader for
each test. This allows us to easily set up different authorizations for
each scenario.

The tests also use a dedicated CollectPermissionScenario which offers a
builder pattern to clarify the scenarios in more details. The build
method decides whether this concerns anonymous access or authenticated.

Some of the scenarios overlap in what edge case they cover. For example,
there are several mismatch scenarios which cover the case where the
resource id does not match any of the available authorizations. Another
example are the scenarios where a union of permissions is formed across
different authorizations. Still, I found it useful to add these as to
cover a wider range of different AuthorizationResourceTypes.

The authEntity method helps easily create AuthorizationEntity objects
where each gets a unique key.